adds the installers athe actual web space builder and the tutorial index

+++ setuplocalwebspace/usr/share/setuplocalwebspace/setuplocalwebspace.sh
@@ -5,10 +5,38 @@
 #   2. if setwebpermissions.sh does not exist copy them in to ~
 #   3. SVN Checkout the tutorials folder
 #   4. Runs setpermissionsforhour.sh 
-cp /usr/share/setuplocalwebspace/setwebpermissions.sh ~/setwebpermissions.sh
-chown $USER:$USER ~/setwebpermissions.sh
-chmod a+x ~/setwebpermissions.sh
-svn checkout http://access-space-web-tutorials.googlecode.com/svn/trunk/tutorials ~/public_html/tutorials
-echo "Start" >> ~/timelog.txt
+
+
+if [ -d /home/$USER/public_html ] 
+then
+	echo "public_html exists"
+else
+	mkdir /home/$USER/public_html
+  if [ -f /home/$USER/public_html/index.html ] || [ -f /home/$USER/public_html/index.php ]
+  then
+    echo "Page Exists"
+  else
+    cat <<END > /home/$USER/public_html/index.php
+<?php
+  header('Location: http://caterpie/~'.get_current_user().'/');
+END
+fi
+
+if [ -f ~/setwebpermissions.sh ] 
+then
+	echo "~/setwebpermissions.sh exists"
+else
+  cp /usr/share/setuplocalwebspace/setwebpermissions.sh ~/setwebpermissions.sh
+  chown $USER:$USER ~/setwebpermissions.sh
+  chmod a+x ~/setwebpermissions.sh
+fi
+
+if [ -d /home/$USER/public_html/tutorials ] 
+then
+	echo "tutorials exists"
+else
+  svn checkout http://access-space-web-tutorials.googlecode.com/svn/trunk/tutorials ~/public_html/tutorials
+fi
+
 /usr/share/setuplocalwebspace/setpermissionsforhour.sh
